+++ test/flop_test.exs
@@ -460,6 +460,26 @@ defmodule FlopTest do
       assert Flop.count(Pet, %Flop{}) == 0
       assert Flop.count(Pet, %Flop{}, prefix: "other_schema") == 1
     end
+
+    test "allows overriding query" do
+      _matching_pets = insert_list(6, :pet, age: 5, name: "A")
+      _more_matching_pets = insert_list(5, :pet, age: 5, name: "B")
+      _non_matching_pets = insert_list(4, :pet, age: 6)
+
+      flop = %Flop{
+        limit: 2,
+        offset: 2,
+        order_by: [:age],
+        filters: [%Filter{field: :age, op: :<=, value: 5}]
+      }
+
+      # default query
+      assert Flop.count(Pet, flop) == 11
+
+      # custom count query
+      assert Flop.count(Pet, flop, count_query: where(Pet, name: "A")) == 6
+      assert Flop.count(Pet, flop, count_query: where(Pet, name: "B")) == 5
+    end
   end
 
   describe "meta/3" do